+++ src/Pinky/Brainfuck/Naive.hs
@@ -1,12 +1,10 @@
-{-# LANGUAGE TypeFamilies #-}
-{-# LANGUAGE FlexibleContexts #-}
-
 module Pinky.Brainfuck.Naive
   ( interpretBasic,
     interpretBasicTape,
   )
 where
 
+import Data.Maybe (fromMaybe)
 import Control.Monad (unless)
 import Data.Foldable (traverse_)
 import Control.Monad.Trans
@@ -17,9 +15,7 @@ import Pinky.Brainfuck.Tape
 -- This is needed because lifting interpretBasic causes the program to
 -- never halt
 interpretBasicTape ::
-  (BrainfuckMachine m, BfCell c, MachineCell m ~ c) =>
-  Bf ->
-  BrainfuckTape c m ()
+  (BfCell c, BrainfuckMachine m c) => Bf -> BrainfuckTape c m ()
 interpretBasicTape = traverse_ interpInstr . _bfCode
   where
     interpInstr Increment = modifyCell (+ 1)
@@ -28,9 +24,8 @@ interpretBasicTape = traverse_ interpInstr . _bfCode
     interpInstr MoveRight = moveHeadRight
     interpInstr Input = do
       char <- lift bfGetChar
-      case char of
-        Just c -> setCell c
-        Nothing -> pure ()
+      -- TODO Abstract out EOF handling
+      setCell (fromMaybe 0 char)
     interpInstr Output = readCell >>= (lift . bfPutChar)
     interpInstr Debug = pure () -- TODO
     interpInstr (Loop code) = interpLoop
@@ -39,6 +34,5 @@ interpretBasicTape = traverse_ interpInstr . _bfCode
           cell <- readCell
           unless (cell == 0) (interpretBasicTape code >> interpLoop)
 
-interpretBasic ::
-  (BrainfuckMachine m, BfCell c, MachineCell m ~ c) => Bf -> m ()
+interpretBasic :: (BfCell c, BrainfuckMachine m c) => Bf -> m ()
 interpretBasic = evalBfTape . interpretBasicTape