+++ src/Pinky/Brainfuck/Optimizer.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE TypeFamilies #-}
-
 module Pinky.Brainfuck.Optimizer
   ( BfOptCell (..),
     interpretOptimized,
@@ -11,6 +9,5 @@ import Pinky.Brainfuck.Optimizer.Internal
 import Pinky.Brainfuck.Machine
 import Pinky.Brainfuck.Tape
 
-interpretOptimized ::
-  (BrainfuckMachine m, BfOptCell c, MachineCell m ~ c) => Bf -> m ()
+interpretOptimized :: (BfOptCell c, BrainfuckMachine m c) => Bf -> m ()
 interpretOptimized = evalBfTape . runBfEffects . runBfOptState . parseRaw