+++ ChangeLog.md
@@ -9,6 +9,16 @@
 ### Changed
 
 - Rename `trimTape` -> `truncateTape` and move into `BrainfuckTape` module
+- Convert `BrainfuckMachine` to use functional dependencies instead of type
+  families. This makes type declarations using it more straight-forward and
+  avoids having to add `{-# LANGUAGE TypeFamilies #-}` everywhere it's used.
+- `BufferMachine` now stores buffers of the cell type rather than strings to
+  get rid of difficulties related to failed character conversions.
+
+### Removed
+
+- `toChar` and `fromChar` from `BfCell`. Character conversion is now the
+  responsability of `BrainfuckMachine` instances.
 
 ## 0.1.0